+++ CHANGELOG.md
@@ -7,14 +7,17 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - Fix potential performance and stability issues from OpenBLAS oversubscription by limiting nested parallelism in [#87](https://github.com/lemma-osu/sklearn-raster/pull/87)
+- Fix slow performance when `nodata_input` contains explicit or inferred missing NoData values
 
 ### Added
 
 - All estimator methods now accept an `inner_thread_limit` keyword argument which limits nested parallelism within Dask workers
+- `nodata_input` values are now validated to ensure that they can be safely cast to the `FeatureArray` data type. If not, a `ValueError` is raised.
 
 ### Changed
 
 - Replaced `utils.features.reshape_to_samples` with the more generalized `utils.decorators.with_inputs_reshaped_to_ndim` which supports reshaping multiple arrays to arbitrary dimensionality
+- `FeatureArray.nodata_input` now uses a masked array to encode missing NoData values, rather than encoding them as `None` and implicitly casting to an `object` type array which impacts ufunc performance.
 
 ## [0.1.0.dev1] - 2025-11-04